More changes made
For our html we added a 'field-error-message' span which will later be used to show what mistakes the user when inputting information. As well as this we generally updated the input and the form area. Then we working on app.js. Created an object that contains all of the dom elements we're going to use. Created event listener for submitting the form alongside functions for validating our form in a custom manner. We use constraint validation api and our own functions to validate the various fields. In the next commit we will start working on 'showRegistrationFormErrors' where we use 'field-error-message' elements to show the errors in the form. For css we added it should be noted that fieldset acts as a flex-box to make things aligned and even. As well as we made main-content a 'row' flex-box just in case we want to add more sections below the login form part.