Finished logic for displaying errors
Created individual functions for each field and displaying the custom messages for those respective fields. Then also created the function for showing all current errors, which is useful when the user presses the submit button. Also created a function for hiding errors, which is useful on hiding the elements when the page loads.