Added additional styling for input fields
We added classes .valid-input and .invalid-input to visually indicate when input fields were valid or invalid. We didn't use :valid and :invalid becuase it was difficult to make that work with the field for retyping in the password. Hard to make 'invalid' trigger the css when the passwords didn't match. So we used did it by adding and removing these classes. Also created functions addValidInputStyle and addInvalidInputStyle, which remove the opposing style and add their own respective styles on the html input elements.