Finished project functionality and requirements
Created a separate file called 'utilities.js' to put less relevant code in. Also moved some of the html element variables out of the DomElements object just to make it so mainly only DOM elements important to the main idea and focus of the idea of the form project are in the object. Also created new function that resets the registration form.